Drop for the character done

+++ server_files/server_shield_factory.cpp
@@ -4,7 +4,7 @@
 #define IRON_SHIELD 8
 
 
-Shield ShieldFactory::make_shield(const int shield_id, Json::Value& config) {
+Shield ShieldFactory::make_shield(const uint8_t shield_id, Json::Value& config) {
     if (shield_id == TURTLE_SHIELD) return TortoiseShield(config);
     else return IronShield(config);
 }
